add metadata and remove claims
