+++ package.json
@@ -2,30 +2,15 @@
   "name": "@boxyhq/saml20",
   "version": "1.0.0",
   "description": "SAML 2.0 token parser for Node.js",
-  "dependencies": {
-    "lodash": "4.17.21",
-    "thumbprint": "0.0.1",
-    "xml-crypto": "2.1.3",
-    "xml2js": "0.4.23",
-    "@xmldom/xmldom": "0.8.2"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/boxyhq/saml20-maintained.git"
-  },
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
-  "files": [
-    "dist"
-  ],
   "keywords": [
     "SAML 2.0",
     "Token Parser"
   ],
-  "scripts": {
-    "build": "tsc -p tsconfig.json",
-    "test": "npm run build && mocha -R spec test"
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/boxyhq/saml20-maintained.git"
   },
+  "license": "MIT",
   "author": "Leandro Boffi (me@leandrob.com)",
   "contributors": [
     {
@@ -37,9 +22,19 @@
       "email": "deepak@boxyhq.com"
     }
   ],
-  "license": "MIT",
-  "readmeFilename": "README.md",
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
+  "files": [
+    "dist"
+  ],
+  "scripts": {
+    "build": "tsc -p tsconfig.json",
+    "sort": "npx sort-package-json",
+    "test": "npm run build && mocha -R spec test"
+  },
   "dependencies": {
+    "@xmldom/xmldom": "0.8.2",
+    "lodash": "4.17.21",
     "rambda": "7.0.3",
     "thumbprint": "0.0.1",
     "xml-crypto": "2.1.3",
@@ -49,5 +44,6 @@
   "devDependencies": {
     "mocha": "9.2.2",
     "typescript": "4.6.3"
-  }
+  },
+  "readmeFilename": "README.md"
 }
